I can only go one way. I've not got a reverse gear

+++ 0x05-pointers_arrays_strings/4-print_rev.c
@@ -11,7 +11,7 @@
 
 void print_rev(char *s)
 {
-	int i = 1;
+	long unsigned int i = 1;
 
 	while (i <= strlen(s))
 	{
